+++ R/lsh.R
@@ -46,8 +46,9 @@
 #'  "\href{http://matthewcasperson.blogspot.com/2013/11/minhash-for-dummies.html}{Minhash
 #'   for Dummies}" (November 14, 2013).
 #'
-#'@seealso \code{\link{minhash_generator}}, \code{\link{lsh_candidates}},
-#'  \code{\link{lsh_query}}, \code{\link{lsh_probability}},
+#'@seealso \code{\link{minhash_generator}}, \code{\link{lsh_add}},
+#'  \code{\link{lsh_candidates}}, \code{\link{lsh_query}},
+#'  \code{\link{lsh_probability}},
 #'  \code{\link{lsh_threshold}}
 #'
 #' @examples
@@ -61,6 +62,36 @@
 #'@export
 lsh <- function(x, bands, progress = interactive()) {
   UseMethod("lsh", x)
+}
+
+#' Add documents to a LSH cache
+#'
+#' This function adds buckets for one or more new documents to an existing
+#' \code{lsh_buckets} object. Use the same \code{bands} value and minhash
+#' function that were used to create the original buckets.
+#'
+#' @param buckets An \code{lsh_buckets} object created by \code{\link{lsh}}.
+#' @param x A \code{\link{TextReuseCorpus}} or
+#'   \code{\link{TextReuseTextDocument}} with minhashes.
+#' @inheritParams lsh
+#' @return An updated \code{lsh_buckets} object.
+#' @seealso \code{\link{lsh}}, \code{\link{lsh_query}},
+#'   \code{\link{lsh_candidates}}
+#' @export
+lsh_add <- function(buckets, x, bands, progress = interactive()) {
+  assert_that(is_lsh_buckets(buckets))
+
+  new_buckets <- lsh(x, bands = bands, progress = progress)
+  new_doc_ids <- unique(new_buckets$doc)
+
+  buckets <- buckets %>%
+    dplyr::filter(!.data$doc %in% new_doc_ids) %>%
+    dplyr::bind_rows(new_buckets) %>%
+    dplyr::arrange(.data$doc)
+
+  class(buckets) <- c("lsh_buckets", setdiff(class(buckets), "lsh_buckets"))
+
+  buckets
 }
 
 #' @export